[#51] The `valueMapping` can ignore the original value
Users could want to ignore the original value when mapping them.

To remains compatible with previous workspace, we add an optional attribute.
The user should specify when to ignore the original value.
In the other case (not configured or configured to false), keep the same logic as before.

This commit only consider the server part.

This commit does the following:

- Add tests about the new functionality
- Add option in typings
- Update `valueMapping` to take into account this new option