[#48] Make the component screen refreshable
When being on the component screen, it is not possible to refresh the page.
The reason is that was lost the information in which workspace we currently are.
To retrieve this information, we need to persist it in the URL so we change the route system.

We only change the `on('navigation')` part having the same impact on the rest of the page to avoid regression.
We also split this part of the code to ease understanding.

This commit does the following:

- In workspaceStore, consider the route for component as `/workspace/:workspaceId/component/:componentId`
- Merge the two `on('navigation')` handlers into one
- Split the navigation handler in several parts but keep global structure (extraction without deep modification)
- The trigger of `navigation_control_done` remains in the global structure (extract from split functions)
- In case of component route, load the workspace and initialize component screen only afterward