[#55] Correct Framacalc component
When using the component framacalc, the component failed each time.
After investigation, the two reasons were:

- The generated URL was invalid with current state of Framacalc (lite.framacalc.org instead of framacalc.org)
- The Framacalc server returned a redirection with given Accept headers

We take the opportunity to be more generic and be able to request any EtherCalc document.
This means changing the notion of `key` into `url`.
However, to avoid breaking existing components, we keep the name `key` and remains compatible in the request.

This commit does the following:

- Update labels in component configuration screen
- Install types of node-fetch to ease IDE integration
- Declare types of input and outputs of Framacalc module
- Use `node-fetch` instead of `http` to simplify request
- Split the job in two parts (`fetchCSV` and `processCSV`) to simplify understanding
- Remove unnecessary code