[#140] Correct status error
When creating a new workspace, the status was not displayed.
The reason was that the workspace creation does not give the exact same model than the list.
It is not really an error in the creation process nor in the list.

To avoid this issue, we force the reload of the whole list of workspaces.

This commit does the following:

* Correct the workspace `create` function on client side to reload the whole list