[#137] Correct login behavior
When the user tries to login, a strange behavior happen.
The user must enter his credentials twice.

After investigation, the reason was that the form was always sent as classic form.
We need to block the default browser behavior for the application to work as expected.

This commit does the following:

* Prevent default browser behavior
* Remove unnecessary validation